Transform to Object Oriented code.
The base FSTreeIntegrityWatch class instances are configuration modules keeping
tool's context such as list of hash algorithms to use and list of files to work
with as well as computed checksums along with their time stamps.

Other modules get FSTreeIntegrityWatch instances as context objects to setup
accordingly, get the settings and save the results.

Objects are constructed using Class::Tiny module. Version 1.001 or newer is
required as older versions lack support of the BUILDARGS method we are using in
some submodules.