Add batch processing.
Operations consisting of two steps on every input file, such as storing
checksums (consisting of the computation of the checksum value and storing the
value to the extended attribute of the file) and verifying checksums (loading
the stored checksum value from the extended attribute and computing verification
checksum from the current file), are done in batches now. Both steps are
performed on all files in the batch prior proceeding to the next batch of files.

The batch processing allows continuous saving of interim results when working on
long lists of input files.

Batch size is configurable on command line.