Implement dump mode storing checksums to dump file only.
In contrast to the save mode saving checksums to the filesystem extended
attributes with optional dump of the results to the integrity database file in
JSON format the dump mode stores the results in the JSON dump only.

Thus, the dump mode is useful on filesystem with no extended attributes support.